requirements.txt
supabase
requests
Pillow
ultralytics
opencv-python
numpy
